Make expectations nestable (#79)
This is done mostly to support further development of be_ok_and and similar
expectations. Also to clean up the code a bit. For now it allows to
nest `have` expectations.

This introduces a breaking change in panic expectations.
`from_panic` and `to_panic` are removed as I doubt they're that useful and
removing them allows to clean up the codebase considerably.


Fixes #78